Oppretter OpprettKlagebehandlingRequest og setter Ytelsestype i tilbakekreving til deprecated
