+++ felles/src/main/kotlin/no/nav/familie/kontrakter/felles/klage/OpprettKlagebehandlingRequest.kt
@@ -1,0 +1,14 @@
+package no.nav.familie.kontrakter.felles.klage
+
+import no.nav.familie.kontrakter.felles.Fagsystem
+import no.nav.familie.kontrakter.felles.Ytelsestype
+import java.time.LocalDate
+
+data class OpprettKlageBehandlingRequest(
+    val ident: String,
+    val ytelsestype: Ytelsestype,
+    val eksternBehandlingId: String,
+    val eksternFagsakId: String,
+    val fagsystem: Fagsystem,
+    val klageMottatt: LocalDate
+)
