+++ felles/src/main/kotlin/no/nav/familie/kontrakter/felles/tilbakekreving/ForhåndsvisVarselbrevRequest.kt
@@ -3,6 +3,7 @@ package no.nav.familie.kontrakter.felles.tilbakekreving
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import no.nav.familie.kontrakter.felles.Fagsystem
 import no.nav.familie.kontrakter.felles.Språkkode
+import no.nav.familie.kontrakter.felles.Ytelsestype
 import java.time.LocalDate
 import javax.validation.constraints.Size
