+++ felles/src/main/kotlin/no/nav/familie/kontrakter/felles/tilbakekreving/HentFagsystemsbehandlingRequest.kt
@@ -1,6 +1,7 @@
 package no.nav.familie.kontrakter.felles.tilbakekreving
 
 import no.nav.familie.kontrakter.felles.Språkkode
+import no.nav.familie.kontrakter.felles.Ytelsestype
 import java.time.LocalDate
 import javax.validation.Valid
 import javax.validation.constraints.Pattern
