+++ felles/src/main/kotlin/no/nav/familie/kontrakter/felles/tilbakekreving/OpprettManueltTilbakekrevingRequest.kt
@@ -1,5 +1,7 @@
 package no.nav.familie.kontrakter.felles.tilbakekreving
 
+import no.nav.familie.kontrakter.felles.Ytelsestype
+
 // Denne brukes fagsystemet for å kalle /api/behandling/manuelt/task/v1 tjeneste i familie-tilbake
 data class OpprettManueltTilbakekrevingRequest(
     val eksternFagsakId: String,
